implement position return from writer
the writer will return a Position object from the record write
this is to allow things that may be listening to the event stream
to take action on collecting data from the writes.

This will be used for creating indexes of the written data
as it it is being writte out.